+++ GameServer/packets/Client/168/HousePermissionsRequestHandler.cs
@@ -55,18 +55,18 @@ namespace DOL.GS.PacketHandler.Client.v168
 
 					pak.WriteByte((byte)level);
 					pak.WriteByte(permission.CanEnterHouse ? (byte)1 : (byte)0);
-					pak.WriteByte(permission.Vault1);
-					pak.WriteByte(permission.Vault2);
-					pak.WriteByte(permission.Vault3);
-					pak.WriteByte(permission.Vault4);
+					pak.WriteByte((byte)permission.Vault1);
+					pak.WriteByte((byte)permission.Vault2);
+					pak.WriteByte((byte)permission.Vault3);
+					pak.WriteByte((byte)permission.Vault4);
 					pak.WriteByte(permission.CanChangeExternalAppearance ? (byte)1 : (byte)0);
-					pak.WriteByte(permission.ChangeInterior);
-					pak.WriteByte(permission.ChangeGarden);
+					pak.WriteByte((byte)permission.ChangeInterior);
+					pak.WriteByte((byte)permission.ChangeGarden);
 					pak.WriteByte(permission.CanBanish ? (byte)1 : (byte)0);
 					pak.WriteByte(permission.CanUseMerchants ? (byte)1 : (byte)0);
 					pak.WriteByte(permission.CanUseTools ? (byte)1 : (byte)0);
 					pak.WriteByte(permission.CanBindInHouse ? (byte)1 : (byte)0);
-					pak.WriteByte(permission.ConsignmentMerchant);
+					pak.WriteByte((byte)permission.ConsignmentMerchant);
 					pak.WriteByte(permission.CanPayRent ? (byte)1 : (byte)0);
 					pak.WriteByte(0x00); // ??
 				}
